Bust URL always on headless check
