+++ lib/plausible/verification/url.ex
@@ -1,0 +1,21 @@
+defmodule Plausible.Verification.URL do
+  def bust_url(url) do
+    cache_invalidator = abs(:erlang.unique_integer())
+    update_url(url, cache_invalidator)
+  end
+
+  defp update_url(url, invalidator) do
+    url
+    |> URI.parse()
+    |> then(fn uri ->
+      updated_query =
+        (uri.query || "")
+        |> URI.decode_query()
+        |> Map.put("plausible_verification", invalidator)
+        |> URI.encode_query()
+
+      struct!(uri, query: updated_query)
+    end)
+    |> to_string()
+  end
+end
